Fix login RLS membership lookup, restore backend tests, and harden auth UX.
Use a dedicated OrganizationMembershipResolver for unscoped membership reads, restore unit/integration tests and CI, and unify frontend auth cache with mock Playwright login.

+++ server/Rental-Manager-Web-App/src/RentalManager.Modules.TenantManagement.Infrastructure/Persistence/Repositories/Common/BaseRepository.cs
@@ -219,7 +219,7 @@ public abstract class BaseRepository<TEntity, TPrimaryKey> :
             FETCH NEXT @PagingLimit ROWS ONLY;
             """;
 
-        parameters.Add("PagingOffset", request.Offset, DbType.Int32);
+        parameters.Add("PagingOffset", request.Offset, DbType.Int64);
         parameters.Add("PagingLimit", request.PageSize, DbType.Int32);
 
         SqlExecution execution = await _executionContext.GetAsync(cancellationToken);
